fix: add eval code for nih
